[FIX] Trader does not care about the net crebit.
The trader must make sure that she only signs debits based on corresponding
credits and that she only considers each credit once. Other than that she
doesn't care about the net effect on her funds.

In fact we MUST not consider the net effect, because she may be both the
incoming and outgoing settlement system for a given transfer, in which
case the net balance change to her on that ledger would be zero.

This is in fact what we ran into in testing and what this commit fixes.