[FEATURE] /quote: add liquidity check
Check liquidity before returning a quote. Error if the balance isn't high
enough. Quoting doesn't actually escrow any funds, so no error is not a guarantee
that a transfer preparation will succeed, and an error is not a guarantee that
it wouldn't.

Fixes #60
Based on https://github.com/interledger/five-bells-connector/pull/61