Respond 200 for all valid notifications.
The connector (the host) should only complain about a notification request if the ledger (client) did something wrong. A notification about a transfer with a bogus memo is not a faulty request, so it should still result in a 2xx response.

In order to ease debugging and testing however, we do return an object that states what the connector thought about the transfer. The ledger will ignore this response, but unit tests and developers may use it to figure out what happened.