refactor: remove ledger-specific code
Move ledger specific code to a separate ledger plugin (`ilp-plugin-bells`). There is only one piece of ledger-specific code remaining after this PR: The info cache still assumes a far-away ledger is a Five Bells ledger. This should be changed, but I couldn't think of a way to do it as a non-breaking change.