feat: BREAKING pass through ILP packet
the connector was stripping the ilp header before sending the last
transfer. the full ilp packet is useful for the recipient to check the
details of the incoming transfer against what was supposed to happen
(i.e. what was in the packet).

this changes the behavior to pass
through the whole ILP packet. this is also useful for when we switch to
binary, as it is arguably simpler for the connector to just forward on
the packet rather than stripping particular fields before sending it to
the receiver. if we do want to strip fields before handing the packet
off to the actual end user, the client libraries can do that.