fix: round the amount by both scale and precision
Ending up with too much precision can happen anytime during normal operation. It should be treated the same as when we end up with too much scale - just round it.

That does mean that for a large payment, let's say a billion dollars, we may end up with a hundred dollar rounding error sometimes, but that's ok. It's still a tiny amount, relatively speaking.