refactor: re-add custom pairs feature
This feature was removed in the previous refactor and is hereby
re-added. While working on this commit I noticed that the feature had no
tests and behaves in an interesting way.

I added a deferred test so we can figure out what behavior we actually
want in the future. (See test/quoteSpec.js)