refactor(peer): split routing peer into sender and receiver
This hopefully makes the code easier to understand. Both the sending
aspect and the receiving aspect of a routing peer has a concept of an
epoch for instance, so it's easy to get those mixed up.

While doing this refactor I actually found and fixed a bug where I was
reading the wrong routing table ID because sender and receiver were on
the same class.