Add `CONNECTOR_ADMIN_API=true` to Quickstart (#488)
Without this additional directive, and because no accounts are defined in this quickstart command, there's nothing on the connector that's listening (no accounts and no admin API) and so there's nothing to keep the node process alive. With the admin API enabled it's running an http server, which will hold the connector process open.